cleanup: extract MessagePipe
